refactor: /utilにutilっぽいのを移動
