+++ src/session.ts
@@ -19,7 +19,7 @@ import { registerTurnOnVideo } from "./sessionHandler/speechTurnOnVideo";
 import { VoiceProvider } from "./speaker/voiceProvider";
 import { createSpeechQueue, SpeechQueue } from "./speechQueue";
 import { SessionEventHandlerRegistrant, SpeechText, VoiceOrStageChannel } from "./types";
-import { createYosugaEmbed } from "./util";
+import { createYosugaEmbed } from "./util/util";
 import { YosugaClient } from "./yosugaClient";
 
 const logger = getLogger("session");
