+++ src/util/result.ts
